+++ nautilus_core/model/src/python/instruments/crypto_future.rs
@@ -111,7 +111,7 @@ impl CryptoFuture {
     #[pyo3(name = "to_dict")]
     fn py_to_dict(&self, py: Python<'_>) -> PyResult<PyObject> {
         let dict = PyDict::new(py);
-        dict.set_item("type", stringify!(CryptoPerpetual))?;
+        dict.set_item("type", stringify!(CryptoFuture))?;
         dict.set_item("id", self.id.to_string())?;
         dict.set_item("raw_symbol", self.raw_symbol.to_string())?;
         dict.set_item("underlying", self.underlying.code.to_string())?;